# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wallet Witness — Nansen-powered blockchain investigation game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CaseSummary {
  id: number;
  title: string;
  description: string;
  difficulty: string;
  chain: string;
  totalClues: number;
  isActive: boolean;
  completedCount: number;
  /** @nullable */
  walletAddressMasked?: string | null;
}

export type ClueDataData = { [key: string]: unknown };

export interface ClueData {
  clueIndex: number;
  clueType: string;
  title: string;
  hint: string;
  /** @nullable */
  nansenEndpoint?: string | null;
  data: ClueDataData;
}

export interface CaseDetail {
  id: number;
  title: string;
  description: string;
  difficulty: string;
  chain: string;
  totalClues: number;
  walletAddressMasked: string;
  narrative: string;
  unlockedClues: ClueData[];
}

export interface ClueInput {
  clueIndex: number;
  sessionId: string;
}

export interface ClueResult {
  clue: ClueData;
  totalNansenCallsMade: number;
  fromCache: boolean;
  nansenConfigured: boolean;
}

export type VerdictInputWalletType = typeof VerdictInputWalletType[keyof typeof VerdictInputWalletType];


export const VerdictInputWalletType = {
  smart_money: 'smart_money',
  whale: 'whale',
  retail: 'retail',
  insider: 'insider',
} as const;

export interface VerdictInput {
  walletType: VerdictInputWalletType;
  /**
     * @minimum 1
     * @maximum 100
     */
  confidence: number;
  sessionId: string;
  cluesUnlocked: number;
}

export type VerdictResultBreakdown = { [key: string]: unknown };

export interface VerdictResult {
  score: number;
  isCorrect: boolean;
  correctType: string;
  correctTypeLabel: string;
  explanation: string;
  walletAddress: string;
  walletAddressMasked: string;
  breakdown: VerdictResultBreakdown;
}

export type ApiStatsCallsByEndpoint = { [key: string]: unknown };

export interface ApiStats {
  totalNansenCalls: number;
  callsByEndpoint: ApiStatsCallsByEndpoint;
  casesCompleted: number;
  nansenConfigured: boolean;
  /** @nullable */
  verdictAccuracy?: number | null;
}

export interface InvestigateInput {
  walletAddress: string;
  chain: string;
}

export type InvestigationResultBalance = { [key: string]: unknown };

export type InvestigationResultPnlSummary = { [key: string]: unknown };

export type InvestigationResultRecentDexTrades = { [key: string]: unknown };

export type InvestigationResultRelatedWallets = { [key: string]: unknown };

export interface InvestigationResult {
  walletAddress: string;
  chain: string;
  balance: InvestigationResultBalance;
  pnlSummary: InvestigationResultPnlSummary;
  recentDexTrades: InvestigationResultRecentDexTrades;
  relatedWallets: InvestigationResultRelatedWallets;
  totalNansenCallsMade: number;
  nansenConfigured: boolean;
}

export interface LeaderboardEntry {
  rank: number;
  score: number;
  walletTypeGuess: string;
  cluesUnlocked: number;
  completedAt: string;
  isCorrect: boolean;
  /** True when this entry belongs to the requesting session */
  isPlayer: boolean;
}

export interface LeaderboardCase {
  caseId: number;
  caseTitle: string;
  difficulty: string;
  entries: LeaderboardEntry[];
  /**
     * The requesting player's best score for this case, or null if they haven't played
     * @nullable
     */
  playerBestScore?: number | null;
}

export interface LeaderboardResponse {
  cases: LeaderboardCase[];
}

export interface SeedInput {
  force?: boolean;
}

export interface SeedResult {
  casesSeeded: number;
  cluesCached: number;
  nansenCallsMade: number;
  message: string;
}

export type GetLeaderboardParams = {
/**
 * Filter to a specific case. If omitted, returns all cases.
 */
caseId?: number;
/**
 * The current player's session ID used to highlight their own entries.
 */
sessionId?: string;
};