# lib/api-zod/src/generated/types/caseDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wallet Witness — Nansen-powered blockchain investigation game API
 * OpenAPI spec version: 0.1.0
 */
import type { ClueData } from './clueData';

export interface CaseDetail {
  id: number;
  title: string;
  description: string;
  difficulty: string;
  chain: string;
  totalClues: number;
  walletAddressMasked: string;
  narrative: string;
  unlockedClues: ClueData[];
}