# lib/api-zod/src/generated/types/caseSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wallet Witness — Nansen-powered blockchain investigation game API
 * OpenAPI spec version: 0.1.0
 */

export interface CaseSummary {
  id: number;
  title: string;
  description: string;
  difficulty: string;
  chain: string;
  totalClues: number;
  isActive: boolean;
  completedCount: number;
  /** @nullable */
  walletAddressMasked?: string | null;
}