# lib/api-zod/src/generated/types/investigationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wallet Witness — Nansen-powered blockchain investigation game API
 * OpenAPI spec version: 0.1.0
 */
import type { InvestigationResultBalance } from './investigationResultBalance';
import type { InvestigationResultPnlSummary } from './investigationResultPnlSummary';
import type { InvestigationResultRecentDexTrades } from './investigationResultRecentDexTrades';
import type { InvestigationResultRelatedWallets } from './investigationResultRelatedWallets';

export interface InvestigationResult {
  walletAddress: string;
  chain: string;
  balance: InvestigationResultBalance;
  pnlSummary: InvestigationResultPnlSummary;
  recentDexTrades: InvestigationResultRecentDexTrades;
  relatedWallets: InvestigationResultRelatedWallets;
  totalNansenCallsMade: number;
  nansenConfigured: boolean;
}