# lib/api-zod/src/generated/types/leaderboardResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wallet Witness — Nansen-powered blockchain investigation game API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardCase } from './leaderboardCase';

export interface LeaderboardResponse {
  cases: LeaderboardCase[];
}